Reorganização dos elementos da página de Perfil
