+++ src/components/CardInfoAddress/CardInfoAddress.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import {InfoAddressContainer,AddressUser,TextAddress} from './styles'
+
+const CardInfoAddress = () => {
+    return (
+        <InfoAddressContainer>
+            <TextAddress>Endereço cadastrado</TextAddress>
+            <AddressUser>Rua almerinda,55 - Ventos Largos</AddressUser>           
+        </InfoAddressContainer>
+    )
+}
+
+export default CardInfoAddress
