+++ src/components/CardInfoAddress/styles.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+export const InfoAddressContainer = styled.div`
+       display:grid;
+       grid-template-rows: 1fr 1fr ;
+       row-gap: 10px; 
+`
+export const AddressUser = styled.div`
+       font-size:1rem;
+       font-weight:normal;
+`
+export const TextAddress = styled.p`
+       color:#777;
+       font-size:1rem;
+       font-weight:normal;
+`
