+++ src/components/CardInfoHistoric/styles.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+export const InfoHistoricContainer = styled.div`
+        p{
+           text-align:center;
+           padding-top:16px;            
+        } 
+`
+export const TextHistoric = styled.div`
+        padding:20px 20px 5px; 
+        font-size:1rem;
+        font-weight:normal;
+`
