+++ src/components/CardInfoUser/CardInfoUser.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import{InfoUserContainer,NameUser,EmailUser,CpfUser} from './styles'
+const CardInfoUser = () => {
+    return (
+        <InfoUserContainer>
+            <NameUser>Laécio da Conceição</NameUser>
+            <EmailUser>llll@gmail.com</EmailUser>
+            <CpfUser>746.766.766-87</CpfUser>
+        </InfoUserContainer>
+    )
+}
+
+export default CardInfoUser
