+++ src/components/CardInfoUser/styles.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+
+export const InfoUserContainer = styled.div`
+       display:grid;
+       grid-template-rows: 1fr 1fr ;
+       row-gap: 10px; 
+`
+export const NameUser = styled.div`
+       font-size:1rem;
+       font-weight:normal;
+`
+export const EmailUser = styled.div`
+       font-size:1rem;
+       font-weight:normal;
+`
+export const CpfUser = styled.div`
+       font-size:1rem;
+       font-weight:normal;
+`
