+++ src/components/PerfilDetail/styles.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components'
+
+export const PerfilDetailContainer = styled.div`
+       
+`
+export const ContainerUser = styled.div`
+       display:flex;
+       justify-content:space-between;
+       padding:20px;
+`
+export const ContainerAddress = styled.div`
+       background-color:#ddd;
+       display:flex;
+       justify-content:space-between;
+       padding:20px;
+       
+`
+export const ContainerHistoric = styled.div`
+       
+`
