+++ src/pages/PerfilPage/styles.js
@@ -3,45 +3,3 @@ import styled from 'styled-components'
 export const ContainerPerfilPage = styled.div`
        font-weight:normal;
 `
-export const ContainerUser = styled.div`
-       display:flex;
-       justify-content:space-between;
-       padding:20px;
-`
-export const ContainerAddress = styled.div`
-       background-color:#ddd;
-       display:flex;
-       justify-content:space-between;
-       padding:20px;
-
-       p{
-         color:#777;
-       }
-`
-export const ContainerHistoric = styled.div`  
-
-       p{
-         text-align:center;
-         padding-top:10px;
-       }       
-`
-export const CardInfo = styled.div`   
-       display:grid;
-       grid-template-rows: 1fr 1fr ;
-       row-gap: 10px; 
-`
-export const NameUser = styled.div`
-
-`
-export const EmailUser = styled.div`
-
-`
-export const CpfUser = styled.div`
-
-`
-export const AddressUser = styled.div`
-       
-`
-export const TextHistoric = styled.div`
-       padding:20px 20px 5px;      
-`
